requirements.txt
python-dotenv
python-telegram-bot
openai
boto3
